% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revise_multiplications.R
\name{revise_multiplication}
\alias{revise_multiplication}
\title{Helper to revise multiplications}
\usage{
revise_multiplication(vn)
}
\arguments{
\item{vn}{numeric: vector of numbers to be used for multiplications}
}
\value{
logical
}
\description{
Helper to revise multiplications
}
