% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revise_multiplications.R
\name{revise_multiplications}
\alias{revise_multiplications}
\title{Aide memoire pour reviser les tables de multiplications}
\usage{
revise_multiplications(tables = c(1:10), repetitions = 12)
}
\arguments{
\item{tables}{numeric: Les tables que l'on veut reviser}

\item{repetitions}{numeric: Le nombre de multiplications dont on doit trouver la reponse}
}
\value{

}
\description{
Aide memoire pour reviser les tables de multiplications
}
